Fixed Canvas with file_exists
Some issues where encountered when no avatar has been uploaded for the
fighters. The Canvas drawing was interrupted and left a blank result.
The issue is now fixed using Php file_exists() function, which is the
only way to find out if an avatar has been picked or not.
If no avatar is found, the fighter is being assigned a mock up image
« Facebook profile » style